moved editableItem to Components and removed mockData file
