+++ client/src/components/EditableItem.vue
