--- server/src/database/mockData.js
@@ -1,381 +1,0 @@
-const products = [
-    {
-        id: "1",
-        picture: `/images/product1.jpg`,
-        title: "Screen",
-        price: 1300,
-        rating: 5
-    },
-    {
-        id: "2",
-        picture: `/images/product2.jpg`,
-        title: "Headset",
-        price: 110,
-        rating: 4
-    },
-    {
-        id: "3",
-        picture: `/images/product3.jpg`,
-        title: "Shampoo",
-        price: 15,
-        rating: 3
-    },
-    {
-        id: "4",
-        picture: `/images/product1.jpg`,
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "5",
-        picture:
-            "https://m.media-amazon.com/images/I/51bq-oNA--L._AC_SR160,160_.jpg",
-        title: "Dress",
-        price: 1500000,
-        rating: 1
-    },
-    {
-        id: "6",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "7",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "8",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "9",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "10",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "11",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "12",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "13",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "14",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "15",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "16",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "17",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "18",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "19",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "20",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "21",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "22",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "23",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "24",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "25",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "26",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "27",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "28",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "29",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "30",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "31",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "32",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "33",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "34",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "35",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "36",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "37",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "38",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "39",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "40",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "41",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "42",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "43",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "44",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "45",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "46",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "47",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "48",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "49",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    },
-    {
-        id: "50",
-        picture: "/images/product1.jpg",
-        title: "Screen2",
-        price: 1200,
-        rating: 2
-    }
-];
-const users = [
-    {
-        email: "mhmdmar@gmail.com",
-        username: "mhmdmar",
-        password: "123456",
-        profilePicture: `/images/logo.png`,
-        cart: {
-            items: [{id: "1", quantity: 5}]
-        }
-    },
-    {
-        email: "admin",
-        username: "admin",
-        password: "admin",
-        cart: {
-            items: [
-                {
-                    id: "2",
-                    quantity: 3
-                },
-                {
-                    id: "3",
-                    quantity: 2
-                }
-            ]
-        }
-    }
-];
